Moved readfile function to common directory
