+++ .vscode/settings.json
@@ -1,5 +1,7 @@
 {
     "files.associations": {
-        "vector": "cpp"
+        "vector": "cpp",
+        "fstream": "cpp",
+        "iosfwd": "cpp"
     }
 }
