+++ day_1/day_1.cpp
@@ -1,24 +1,7 @@
 #include <iostream>
-#include <fstream>
-#include <vector>
+#include "../common/readFile.cpp"
 
 using namespace std;
-
-vector<int> readFile() {
-    vector<int> data;
-
-    ifstream in("day_1_input.txt", ios::in);
-
-    int temp;
-
-    while(in >> temp){
-        data.push_back(temp);
-    }
-
-    in.close();
-    
-    return data;
-}
 
 int stageOne() {
     vector<int> data = readFile();
