+++ day_1/makefile
@@ -1,0 +1,2 @@
+all:
+	g++ day_1.cpp -o day_1
